requirements.txt
cloudpickle==1.6.0
h5py==3.1.0
keras==2.4.3
lenden==0.0.1.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.2.5
protobuf==3.20.1
pynrrd==0.4.2
python-decouple==3.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.0
simpleitk==2.1.1
tensorflow==2.5.0
Unipath==1.1
